requirements.txt
# Python packages for SolidWorks automation
pywin32>=305
comtypes>=1.4.0
pythonnet>=3.0.0
openpyxl>=3.1.0

# AI Integration
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# GitHub Integration (for PR review bot)
PyGithub>=2.1.1
gitpython>=3.1.40

